+++ ChangeLog.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### added
 
+- added anonymous function support an optional name. e.g. `(fn double [x] (* 2 x))`
 - added function _doc_
 - added function _list*_
 - added function _io/file-size_